chore: move location of trip details utils
