+++ src/pages/TripBookingPage/TripBookingPage.tsx
@@ -4,10 +4,9 @@ import LongArrow from '../../components/icons/LongArrow';
 import { TripCardAccordion } from '../../components';
 import { ITrip, ISeatType, IBusType } from '../../types';
 import { useAppSelector } from '../../app/hooks';
-import { fromSerializable } from '../../app/features/utils/tripDetailsHelperFns';
+import { fromSerializable } from '../../utils/tripDetailsUtils';
 
-
-// TODO: fetch data from store 
+// TODO: fetch data from store
 const dummyTripData: ITrip = {
     id: '1',
     origin: {
@@ -30,7 +29,9 @@ const dummyTripData: ITrip = {
 };
 
 export const TripBookingPage = () => {
-    const state = useAppSelector((state) => fromSerializable(state.tripDetails));
+    const state = useAppSelector((state) =>
+        fromSerializable(state.tripDetails)
+    );
     return (
         <>
             <Stack direction="row" alignItems="center" mt={4} mb={3} gap="2rem">
@@ -42,7 +43,11 @@ export const TripBookingPage = () => {
                     {dummyTripData.destination.name}
                 </Typography>
             </Stack>
-            <TripCardAccordion defaultExpanded={true} data={state} mode='view'/>
+            <TripCardAccordion
+                defaultExpanded={true}
+                data={state}
+                mode="view"
+            />
             <PassengerDetailsForm />
         </>
     );
